requirements.txt
torch>=2.0.0
torchvision>=0.15.0
mlflow==2.9.0
dvc==3.30.1
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
Pillow>=10.0.0
numpy>=1.24.0
pytest
httpx
matplotlib
seaborn
scikit-learn
